+++ src/_types/YobtaOptionalIfUnkown.ts
@@ -1,0 +1,1 @@
+export type YobtaOptionalIfUnkown<I, O> = I extends unknown ? O | undefined : O