+++ src/_types/YobtaPath.ts
@@ -1,0 +1,1 @@
+export type YobtaPath = (number | string)[]