+++ src/ruleYobta/index.ts
@@ -1,14 +1,10 @@
-import type { YobtaContext } from '../_internal/createContext/index.js'
+import type { YobtaContext } from '../_types/YobtaContext'
+import type { YobtaOptionalIfUnkown } from '../_types/YobtaOptionalIfUnkown'
 
 // #region SyncRules
 type SyncValidator<I, O> = (input: I, context: YobtaContext) => O
-export type OptionalIfUnkown<I, O> = I extends unknown ? O | undefined : O
 
 export type SyncRule<I, O> = (context: YobtaContext) => (input: I) => O
-
-export type OptionalSyncRule<I, O> = (
-  context: YobtaContext,
-) => (input: I) => OptionalIfUnkown<I, O>
 
 export type SyncRules = [AnySyncRule, ...AnySyncRule[]]
 export type AnySyncRule = SyncRule<any, any>
@@ -57,7 +53,7 @@ type AsyncValidator<I, O> = (input: I, context: YobtaContext) => Promise<O>
 
 export type AsyncRule<I, O> = (
   context: YobtaContext,
-) => (input: I) => Promise<OptionalIfUnkown<I, O>>
+) => (input: I) => Promise<YobtaOptionalIfUnkown<I, O>>
 
 export type AnyAsyncRule = AsyncRule<any, any>
 export type AnySyncOrAsyncRule = AnyAsyncRule | AnySyncRule